Move prefix event to task and task list classes instead
