+++ src/Features/OnboardingTasks/TaskList.php
@@ -251,6 +251,16 @@ class TaskList {
 	}
 
 	/**
+	 * Prefix event for track event naming.
+	 *
+	 * @param string $event_name Event name.
+	 * @return string
+	 */
+	public function prefix_event( $event_name ) {
+		return $this::get_list_id() . '_tasklist_' . $event_name;
+	}
+
+	/**
 	 * Get the list for use in JSON.
 	 *
 	 * @return array
